+++ R/error-checks.R
@@ -18,3 +18,18 @@
   }
   return(invisible(TRUE))
 }
+
+#' Check if the `object` contains active binding with the name `field`
+#'
+#' @param object A class or an instance of a class to check
+#' @param field Name of the field
+#'
+#' @return `TRUE` if the `object` has an active binding `field`,
+#' `FALSE?  otherwise.
+#' @keywords internal
+.validateClassHasField <- function(object, field){
+  if (!field %in% names(object)){
+    return(FALSE)
+  }
+  return(TRUE)
+}