+++ R/messages.R
@@ -161,3 +161,9 @@ messages$errorWrongArguments <- function(expectedArguments) {
     expectedArguments, "."
   )
 }
+
+messages$invalidConfigurationPropertyFromExcel <- function(propertyName, configurationType){
+  paste0("Trying to apply property '", propertyName, "' that is not supported by
+         the configuration '", configurationType, "'! Check column names in the
+         excel file defining plot configurations.")
+}